codecheck

+++ rosen/modules/render_service/core/pipeline/rs_uni_ui_capture.cpp
@@ -112,7 +112,7 @@ std::shared_ptr<Media::PixelMap> RSUniUICapture::TakeLocalCapture()
 bool RSUniUICapture::CopyDataToPixelMap(std::shared_ptr<Drawing::Image> img,
     std::shared_ptr<Media::PixelMap> pixelmap)
 {
-    auto size = pixelmap->GetRowBytes() * pixelmap->GetHeight();
+    auto size = static_cast<int64_t>(pixelmap->GetRowBytes()) * pixelmap->GetHeight();
     if (size <= 0) {
         return false;
     }
